feat: 为各页面设置独立的浏览器标题
在 layout.tsx 配置标题模板（`%s | movieclaw console`），
首页（Overview）和健康检查页（Health Check）分别导出各自的 metadata.title。
因 health/page.tsx 原为客户端组件无法导出 metadata，
将交互逻辑提取至 HealthClient.tsx，页面文件改为服务端组件负责元数据声明。
